:zap: Fix issues on FTP-Node
