+++ packages/nodes-base/package.json
@@ -210,7 +210,7 @@
       "dist/nodes/ExecuteWorkflow.node.js",
       "dist/nodes/Facebook/FacebookGraphApi.node.js",
       "dist/nodes/FileMaker/FileMaker.node.js",
-      "dist/nodes/FileTransfer.node.js",
+      "dist/nodes/Ftp.node.js",
       "dist/nodes/Freshdesk/Freshdesk.node.js",
       "dist/nodes/Flow/Flow.node.js",
       "dist/nodes/Flow/FlowTrigger.node.js",
